main,watcher: set up callbacks for window creation/destruction
This commit also introduces the structure I envision for running this
program, i.e. by running goroutines in an errgroup. The X event loop
is started by the `main` function and not the `Watcher`, since the
`main` function is the original "owner" of the X connection and should
generally be tasked with starting everything up.